Added LED blinking
New task for blinking an LED has been added: it is used to check, when
the I2C task is stuck if the rest of the CPU is still running.

Improved the I2C scan to make it faster (and to push the pus at full
speed during the scan).